moved code to app.js
